feat(client): 알림 바로가기 버튼 안내 툴팁 추가 (#246)
